requirements.txt
# === Core ===
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-telegram-bot>=21.9
openai>=1.59.0
httpx>=0.28.0
pydantic>=2.10.0
python-dotenv>=1.0.0

# === Retry / resilience ===
tenacity>=9.0.0

# === Charts ===
matplotlib>=3.10.0
numpy>=2.2.0

# === Dev / Testing ===
pytest>=8.3.0
pytest-asyncio>=0.25.0
ruff>=0.9.0
mypy>=1.14.0

# === Optional: Google Sheets ===
# gspread>=6.1.0
# google-auth>=2.37.0
